Move Status to util package
